add parameter uniquification style
Adds a new module uniquification style that incorporates the parameter
names and values into the filename, instead of a sequentially increasing
number (the traditional approach). This is useful when using Genesis as
part of an ASIC flow, instead of just as a tool for design-space
exploration. By having the paramter names/values as part of the file
name, it makes it easier to identify which instance of a module
corresponds to which param set.

The parameter uniquification style is the new default.

Parameter names are abbreviated before being included in the file name.
The abbreviation logic prefers to take only the first letter from each
word, together with full numbers. However, if two parameters produce the
same abbreviation, then the abbrev alg will increase the number of
letters in order to create a unique abbreviation.

For example;
 - A single parameter "MY_PARAM" is abbreviated to "MP".
 - Two parameters "MY_PARAM" and "MIN_POS" both produce "MP" if we only
   take the first letter of each word. Instead, the algo will produce
   "MYP" and "MIP" respectively for the two parameters.

The values are converted to strings with a little sanitization (periods
and quotes are converted to underscores). This works well for simple
data types (e.g., strings, numbers), but produces unusual results for
complex data types and references.