Parameter uniquification style (+ other small improvements) (#19)
* cache instance path

Cache the instance path to speed up calls to get_instance_path. This function
is called frequently, so cache the result to avoid repeated string
concatenation.

* add parameter uniquification style

Adds a new module uniquification style that incorporates the parameter
names and values into the filename, instead of a sequentially increasing
number (the traditional approach). This is useful when using Genesis as
part of an ASIC flow, instead of just as a tool for design-space
exploration. By having the paramter names/values as part of the file
name, it makes it easier to identify which instance of a module
corresponds to which param set.

The parameter uniquification style is the new default.

Parameter names are abbreviated before being included in the file name.
The abbreviation logic prefers to take only the first letter from each
word, together with full numbers. However, if two parameters produce the
same abbreviation, then the abbrev alg will increase the number of
letters in order to create a unique abbreviation.

For example;
 - A single parameter "MY_PARAM" is abbreviated to "MP".
 - Two parameters "MY_PARAM" and "MIN_POS" both produce "MP" if we only
   take the first letter of each word. Instead, the algo will produce
   "MYP" and "MIP" respectively for the two parameters.

The values are converted to strings with a little sanitization (periods
and quotes are converted to underscores). This works well for simple
data types (e.g., strings, numbers), but produces unusual results for
complex data types and references.

* document unqstyle option

Add description of the -unqstyle parameter.

* allow richer inputlist files

Allow inputlist files to include these commands, in addition to files:
 -input
 -inputlist
 -incpath
 -srcpath

* add generate_w_name function

Add generate_w_name to generate a module with a specific name.

* add parse_unprocessed_file

Add parse_unprocessed_file to parse a file if it hasn't already been
parsed. Based on parse_files.

* add find_file_safe

Add a version of find_file that does _not_ error out when a file is not
found. The function returns undef if the file is not found.

* add_suffix func / parse files in load_base_module

Add add_suffix function that attempts to find the suffix to add to a
base module name.

Call add_suffix and parse_unprocess_file from load_base_module.

* extract common code into func

Extract common code from parse_files and parse_unprocessed_file.

* avoid duplicate parsing

* support directories in filename for find_file_safe

Handle the case where the $file variable contains a path in
find_file_safe.

* fix build problems with test/glctest

Handle the case that files with paths are included on the command line.
Need to remove the paths from their names when caching.  Look in the
DependHistogram when trying to find a file, before invoking
find_file_safe.

* Update version info for the first time in twelve years(!!!)
* do not print debug info to verilog files as verilog code!!!
* hand merge w status-report-sort
* restore a gold file to working version
* Update test/README
* minor mods to gleng/param_uniquify-2; all tests pass now, with no redundancy in generated files
---------

Co-authored-by: steveri <steveri@steveri.com>